refactor: extract hugging face text2img as component
